Refactor minimal apis into controllers as it was getting out of hand
