+++ LiftLog.Api/Controllers/UserController.cs
@@ -1,0 +1,122 @@
+using FluentValidation;
+using LiftLog.Api.Db;
+using LiftLog.Api.Models;
+using LiftLog.Api.Service;
+using LiftLog.Lib.Models;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+
+namespace LiftLog.Api.Controllers;
+
+[ApiController]
+[Route("[controller]")]
+public class UserController(UserDataContext db, PasswordService passwordService) : ControllerBase
+{
+    [Route("/create")]
+    [HttpPost]
+    public async Task<IActionResult> CreateUser(
+        CreateUserRequest request,
+        [FromServices] IValidator<CreateUserRequest> validator
+    )
+    {
+        var validationResult = await validator.ValidateAsync(request);
+        if (!validationResult.IsValid)
+        {
+            return BadRequest(validationResult.Errors);
+        }
+        if (await db.Users.AnyAsync(x => x.Id == request.Id))
+        {
+            return BadRequest(new string[] { "User already exists" });
+        }
+        var password = Guid.NewGuid().ToString();
+        var hashedPassword = passwordService.HashPassword(password, out var salt);
+        var user = new User
+        {
+            Id = request.Id,
+            HashedPassword = hashedPassword,
+            Salt = salt,
+            LastAccessed = DateTimeOffset.UtcNow,
+            EncryptionIV = [],
+        };
+
+        await db.Users.AddAsync(user);
+        await db.SaveChangesAsync();
+        return Ok(new CreateUserResponse(password));
+    }
+
+    [Route("{id}")]
+    [HttpGet]
+    public async Task<IActionResult> GetUser(Guid id)
+    {
+        var user = await db.Users.FindAsync(id);
+        if (user == null)
+        {
+            return NotFound();
+        }
+        user.LastAccessed = DateTimeOffset.UtcNow;
+        await db.SaveChangesAsync();
+        return Ok(
+            new GetUserResponse(
+                EncryptedCurrentPlan: user.EncryptedCurrentPlan,
+                EncryptedProfilePicture: user.EncryptedProfilePicture,
+                EncryptedName: user.EncryptedName,
+                EncryptionIV: user.EncryptionIV
+            )
+        );
+    }
+
+    [Route("delete")]
+    [HttpPost]
+    public async Task<IActionResult> DeleteUser(
+        DeleteUserRequest request,
+        [FromServices] IValidator<DeleteUserRequest> validator
+    )
+    {
+        var validationResult = await validator.ValidateAsync(request);
+        if (!validationResult.IsValid)
+        {
+            return BadRequest(validationResult.Errors);
+        }
+        var user = await db.Users.FindAsync(request.Id);
+        if (user == null)
+        {
+            return NotFound();
+        }
+        if (!passwordService.VerifyPassword(request.Password, user.HashedPassword, user.Salt))
+        {
+            return Unauthorized();
+        }
+        db.Users.Remove(user);
+        await db.SaveChangesAsync();
+        return Ok();
+    }
+
+    [Route("/")]
+    [HttpPut]
+    public async Task<IActionResult> PutUser(
+        PutUserDataRequest request,
+        [FromServices] IValidator<PutUserDataRequest> validator
+    )
+    {
+        var validationResult = await validator.ValidateAsync(request);
+        if (!validationResult.IsValid)
+        {
+            return BadRequest(validationResult.Errors);
+        }
+        var user = await db.Users.FindAsync(request.Id);
+        if (user == null)
+        {
+            return NotFound();
+        }
+        if (!passwordService.VerifyPassword(request.Password, user.HashedPassword, user.Salt))
+        {
+            return Unauthorized();
+        }
+        user.EncryptedCurrentPlan = request.EncryptedCurrentPlan;
+        user.EncryptedProfilePicture = request.EncryptedProfilePicture;
+        user.EncryptedName = request.EncryptedName;
+        user.EncryptionIV = request.EncryptionIV;
+        await db.SaveChangesAsync();
+        return Ok();
+    }
+}
